add Inject annotation in ImageFileManager.kt and update generateImageUri function
