+++ app/src/main/java/com/guru/recognizetext/utils/ImageFileManager.kt
@@ -1,0 +1,30 @@
+package com.guru.recognizetext.utils
+
+import android.content.Context
+import android.net.Uri
+import android.os.Environment
+import androidx.core.content.FileProvider
+import java.io.File
+import javax.inject.Inject
+
+class ImageFileManager @Inject constructor(private val context: Context ) {
+    private lateinit var tempFile: File
+     lateinit var imageUri: Uri
+
+    fun createTempFile() {
+        tempFile = File.createTempFile(
+            "image",
+            ".jpg",
+            context.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
+        )
+        generateImageUri()
+    }
+
+  private fun generateImageUri() {
+        imageUri = FileProvider.getUriForFile(
+            context,
+            context.packageName + ".fileprovider",
+            tempFile
+        )
+    }
+}
